move filter functions to filter.go with tests + getFilterSortText
