+++ go.mod
@@ -3,10 +3,10 @@ module github.com/ONSdigital/dp-frontend-search-controller
 go 1.14
 
 require (
-	github.com/ONSdigital/dp-api-clients-go v1.30.1
-	github.com/ONSdigital/dp-frontend-models v1.6.2
+	github.com/ONSdigital/dp-api-clients-go v1.32.4
+	github.com/ONSdigital/dp-frontend-models v1.9.2
 	github.com/ONSdigital/dp-healthcheck v1.0.5
-	github.com/ONSdigital/go-ns v0.0.0-20200511161740-afc39066ee62
+	github.com/ONSdigital/go-ns v0.0.0-20200902154605-290c8b5ba5eb
 	github.com/ONSdigital/log.go v1.0.1
 	github.com/gorilla/mux v1.8.0
 	github.com/kelseyhightower/envconfig v1.4.0
