+++ handlers/handlers.go
@@ -6,12 +6,9 @@ import (
 	"errors"
 	"net/http"
 	"net/url"
-	"strings"
-
-	"github.com/ONSdigital/dp-frontend-search-controller/data"
-	"github.com/ONSdigital/dp-frontend-search-controller/mapper"
 
 	search "github.com/ONSdigital/dp-api-clients-go/site-search"
+	"github.com/ONSdigital/dp-frontend-search-controller/mapper"
 	"github.com/ONSdigital/log.go/log"
 )
 
@@ -112,73 +109,4 @@ func read(w http.ResponseWriter, req *http.Request, rendC RenderClient, searchC 
 		log.Event(ctx, "getting search page failed", log.Error(err), log.ERROR)
 	}
 	return
-}
-
-func mapFilterTypes(ctx context.Context, query url.Values) (apiQuery url.Values, err error) {
-	apiQuery, err = url.ParseQuery(query.Encode())
-	if err != nil {
-		log.Event(ctx, "failed to parse copy of query for mapping filter types", log.Error(err), log.ERROR)
-		return nil, err
-	}
-	filters := apiQuery["filter"]
-	if len(filters) > 0 {
-		var newFilters []string
-		for _, fType := range filters {
-			found := false
-			for _, category := range data.Category {
-				for _, searchType := range category {
-					if fType == searchType.QueryType {
-						found = true
-						newFilters = append(newFilters, searchType.SubTypes)
-						break
-					}
-				}
-				if found {
-					break
-				}
-			}
-			if !found {
-				return nil, errFilterType
-			}
-		}
-		apiQuery.Del("filter")
-		apiQuery.Set("content_type", strings.Join(newFilters, ","))
-	}
-	return apiQuery, nil
-}
-
-func mapCountFilterTypes(ctx context.Context, apiQuery url.Values, searchC SearchClient) (mappedContentType []search.ContentType, err error) {
-	//Remove filter to get count of all types for the query from the client
-	apiQuery.Del("content_type")
-	countResp, err := searchC.GetSearch(ctx, apiQuery)
-	if err != nil {
-		log.Event(ctx, "getting search query count from client failed", log.Error(err), log.ERROR)
-		return nil, err
-	}
-	countFilter := make(map[string]int)
-	for _, contentType := range countResp.ContentTypes {
-		foundFilter := false
-		for _, category := range data.Category {
-			for _, searchType := range category {
-				mapfilters := strings.Split(searchType.SubTypes, ",")
-				for _, filter := range mapfilters {
-					if filter == contentType.Type {
-						countFilter[searchType.QueryType] += contentType.Count
-						foundFilter = true
-					}
-				}
-			}
-		}
-		if !foundFilter {
-			return nil, errors.New("filter type from client not available in filterTypes map")
-		}
-	}
-	for k, v := range countFilter {
-		mappedContentType = append(mappedContentType, search.ContentType{
-			Type:  k,
-			Count: v,
-		})
-	}
-
-	return mappedContentType, nil
 }
